-moved Input styles + variables to teams theme
