+++ packages/react/src/themes/base/componentVariables.ts
@@ -3,8 +3,6 @@ export { default as FlexItem } from './components/Flex/flexItemVariables'
 
 export { default as Icon } from './components/Icon/iconVariables'
 
-export { default as Input } from './components/Input/inputVariables'
-
 export { default as Loader } from './components/Loader/loaderVariables'
 
 export { default as Grid } from './components/Grid/gridVariables'
