+++ packages/react/src/themes/teams/componentStyles.ts
@@ -48,6 +48,8 @@ export { default as Icon } from './components/Icon/iconStyles'
 
 export { default as Image } from './components/Image/imageStyles'
 
+export { default as Input } from './components/Input/inputStyles'
+
 export { default as Label } from './components/Label/labelStyles'
 
 export { default as Layout } from './components/Layout/layoutStyles'
